Move markstyle impl into submodule
